requirements.txt
python-dotenv~=1.0.1
langchain~=0.3.18
langchain-openai~=0.3.4